pyproject.toml
[tool.poetry]
name = "__auth2.0"
version = "0.1.0"
description = ""
authors = ["Do Thanh Long <longdt@hybrid-technologies.vn>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
drf-social-oauth2 = "^1.0.9"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
